fix(fake): valor float integral aceito, mensagem específica e metadata validada no feedback
Fix M1 da revisão final ("Logs idênticos à Portkey"): `valor: float`
→ `int` (o tipo que o núcleo de fato aceita, RN-FE-01/migration 065).
O fake recusava 4.0 (isinstance(valor, int) é False para float em
Python) enquanto o núcleo aceitaria — JSON manda 4.0, o núcleo (JS)
recebe o número 4 e Number.isInteger(4) passa. `valor.is_integer()`
alinha os dois sem abrir mão de recusar fracionário (4.5).

Fix I6: mensagem ESPECÍFICA para valor_invalido/peso_invalido,
espelhando o 3º argumento de politicaDeForma que o núcleo agora
passa — não mais a genérica. log_id_invalido/metadata_invalida
continuam genéricos.

Fix M2: o fake não validava `metadata` nenhuma (o núcleo valida via
validarMetadataDefault: objeto de strings, até 128 caracteres por
valor, 400 metadata_invalida) e a resposta 201 não ecoava o campo (o
núcleo ecoa).

+++ src/avalonops/__init__.py
@@ -111,7 +111,11 @@ class _Feedback:
         self,
         *,
         request_id: str,
-        valor: float,
+        # Fix M1 da revisão final ("Logs idênticos à Portkey"): `valor` é
+        # INTEIRO na escala da Portkey (-10..10, RN-FE-01/migration 065,
+        # governanca-plataforma/src/feedback/validacao.ts) — `float` aqui
+        # divergia do próprio tipo que o núcleo aceita.
+        valor: int,
         peso: float | None = None,
         metadata: Mapping[str, str] | None = None,
     ) -> Any:
